[FIX] change the module of the override of name_search
